Remove stylint, use modern eslint.
I've not tried to maintain the exact rules but the default set is laxer except
a couple that I've relaxed.

Stylint is unmaintained and contributed to the many critical audit issues.

From a quick look there does not seem to be a well maintained linter for Stylus
these days.

We're now down to a few high issues that aren't of concern and more importantly
the output is small enough to comprehend.